third_party/blink/web_tests/external/wpt/css/css-mixins/invalidation.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] invalidation of @mixin from same stylesheet
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
[FAIL] invalidation of @mixin from different stylesheet
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
